my_site/requirements.txt
psycopg2-binary==2.9.7
Pillow==10.0.1
gunicorn==21.2.0
django-ckeditor==6.7.3